fix line numbers interaction with code context
CPython commit 213ce1 turned IDLE built-in extensions into regular
features. This breaks the LineNumbers extension in its attempt to
interact with the Code-Context feature. The fix provided by this
changeset is simply to move line numbers to the right of the source
code, which is a somewhat ununsual choice, but eliminates the need
to interact with code context entirely.

This changeset also makes the line numbers respect the configured
color scheme.